Find "best" variogram parameters.

Find the "best" variogram parameters for a given geodata object. Several different models are tested; the one with the lowest least squares error is chosen.

krig_auto_params(geodata, trend = "cte", breaks = krig_breaks(2, 320,
  30))

Arguments

geodata
a list containing element coords as described next. Typically an object of the class "geodata" - a geoR data-set. If not provided the arguments coords must be provided instead.

trend
specifies the mean part of the model. See documentation of trend.spatial for further details. Defaults to "cte".

breaks
a vector with values to define the variogram binning. Only used when option = "bin". See DETAILS below for more details on how to specify the bins.

Value

A list of the best fitted variogram parameters. The following description is adapted from geoR::variog() -- which you should see for more details:

nugget: value of the nugget parameter. An estimated value if fix.nugget = FALSE or a fixed value if fix.nugget = TRUE.

sill and range: First and second elements of cov.pars` -- a two elements vector with estimated values of the covariance parameters sigma^2 and phi, respectively.

kappa: Fixed value of the smoothness parameter.

model: Name of the correlation function (see cov.model).

minVM: The minimum fit error.

vg: The variogram.

Breaks default

The default breaks argument is set to have more points where the exponential curve rises the most and fewer at large distances. This means that the curve fitting is not overly biased by points beyond the effective maximum range.

See also